Restructed code project
